requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.4
pyjwt==2.8.0